@@ -15,7 +15,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -23,16 +23,15 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  bip39: ^1.0.6
+  alan: ^0.40.1+1
   clipboard: ^0.1.3
   cupertino_icons: ^0.1.3
   equatable: ^2.0.0
   ffi: ^1.0.0
   flutter_bloc: ^7.0.0
-  flutter_svg: ^0.19.3
+  flutter_svg: ^0.21.0
   intl: ^0.17.0
-  wallet:
-    path: packages/wallet
+  pedantic: ^1.9.0
 
 dev_dependencies:
   flutter_test:
