@@ -1,4 +1,4 @@
 export 'account_generated/export.dart';
 export 'generate_account/export.dart';
 export 'import_mnemonic/export.dart';
-export 'home.dart';
+export 'home/export.dart';
