@@ -4,7 +4,7 @@ import 'package:dam/ui/export.dart';
 /// Contains all the Desmos text styles used through all the application.
 class DesmosTextStyles {
   static TextStyle title(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: Colors.black,
           fontSize: 20,
           fontWeight: FontWeight.w500,
@@ -12,23 +12,23 @@ class DesmosTextStyles {
   }
 
   static TextStyle thinHeader(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: Colors.white,
-          fontSize: 18,
+          fontSize: DesmosPlatform.isMobile ? 18 : 24,
           fontWeight: FontWeight.w400,
         );
   }
 
   static TextStyle largeBody(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: Colors.black,
-          fontSize: 16,
+          fontSize: DesmosPlatform.isMobile ? 16 : 20,
           fontWeight: FontWeight.w400,
         );
   }
 
   static TextStyle smallBody(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: DesmosColors.grey,
           fontSize: 12,
           fontWeight: FontWeight.w400,
@@ -36,7 +36,7 @@ class DesmosTextStyles {
   }
 
   static TextStyle thinBodyGrey(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: DesmosColors.grey,
           fontSize: 14,
           fontWeight: FontWeight.w400,
@@ -44,7 +44,7 @@ class DesmosTextStyles {
   }
 
   static TextStyle thinBodyBlack(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: Colors.black,
           fontSize: 14,
           fontWeight: FontWeight.w400,
@@ -52,7 +52,7 @@ class DesmosTextStyles {
   }
 
   static TextStyle boldBody(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: Colors.black,
           fontSize: 14,
           fontWeight: FontWeight.w500,
@@ -60,7 +60,7 @@ class DesmosTextStyles {
   }
 
   static TextStyle secondaryButton(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: DesmosColors.blue,
           fontSize: 14,
           fontWeight: FontWeight.w400,
@@ -68,7 +68,7 @@ class DesmosTextStyles {
   }
 
   static TextStyle error(BuildContext context) {
-    return Theme.of(context).textTheme.bodyText1.copyWith(
+    return Theme.of(context).textTheme.bodyText1!.copyWith(
           color: Color(0xFFFF0404),
           fontSize: 16,
           fontWeight: FontWeight.w400,
